ai/ai_server/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.4
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
PyMySQL==1.1.1
python-decouple==3.8
regex==2025.7.34
requests==2.32.4
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xlwt==1.3.0
